FileManagement: Refuse to close the rootfs fd
If the guest tries to close the rootfs fd, ignore it. This is usually
over-eager fd cleanup code before exec. Since the rootfs fd is
O_CLOEXEC, it will be closed on execve anyway.

Fixes the Proton wrapper closing the rootfs fd, which was harmless since
the execve codepath didn't use it, but becomes a problem after the
subsequent patch changes GetEmulatedPath to call GetEmulatedFDPath.

Signed-off-by: Asahi Lina <lina@asahilina.net>